refactor: replace `~~` imports with `#` alias for server and shared modules
